move more simulation stuff out of acquire
