+++ PYME/simulation/fluorophores.py
