+++ PYME/simulation/illuminate.pyx
